Rectified update product method

+++ ProductService/src/main/java/com/example/ProductService/Services/fakeStoresProductService.java
@@ -30,7 +30,7 @@ public class fakeStoresProductService implements ProductService{
     }
 
     @Override
-    public Product createProduct(String name, String description, String category, int price) {
+    public Product createProduct(String name, String description, String category, Double price) {
         return null;
     }
